requirements.txt
lxml==6.0.2
numpy==2.4.1
pandas==3.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
scipy==1.17.0
six==1.17.0
spicy==0.16.0
tzdata==2025.3
